+++ core/collections/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.0.0](https://github.com/nativecode-dev/nofrills/compare/@nofrills/collections@3.8.7...@nofrills/collections@4.0.0) (2019-06-22)
+
+
+### Features
+
+* **lincoln:** debug and console no longer re-exports their dependency ([0101a8e](https://github.com/nativecode-dev/nofrills/commit/0101a8e))
+
+
+### BREAKING CHANGES
+
+* **lincoln:** must now have an explicit dependency on @nofrills/lincoln
+
+
+
+
+
+## [3.8.7](https://github.com/nativecode-dev/nofrills/compare/@nofrills/collections@3.8.6...@nofrills/collections@3.8.7) (2019-06-22)
+
+**Note:** Version bump only for package @nofrills/collections
+
+
+
+
+
+## [3.8.6](https://github.com/nativecode-dev/nofrills/compare/@nofrills/collections@3.8.5...@nofrills/collections@3.8.6) (2019-06-22)
+
+**Note:** Version bump only for package @nofrills/collections
+
+
+
+
+
 ## [3.8.5](https://github.com/nativecode-dev/nofrills/compare/@nofrills/collections@3.8.2...@nofrills/collections@3.8.5) (2019-06-22)
 
 **Note:** Version bump only for package @nofrills/collections