+++ packages/projector/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.0.0](https://github.com/nativecode-dev/nofrills/compare/@nofrills/projector@4.8.12...@nofrills/projector@5.0.0) (2019-06-22)
+
+
+### Features
+
+* **lincoln:** debug and console no longer re-exports their dependency ([0101a8e](https://github.com/nativecode-dev/nofrills/commit/0101a8e))
+
+
+### BREAKING CHANGES
+
+* **lincoln:** must now have an explicit dependency on @nofrills/lincoln
+
+
+
+
+
+## [4.8.12](https://github.com/nativecode-dev/nofrills/compare/@nofrills/projector@4.8.11...@nofrills/projector@4.8.12) (2019-06-22)
+
+**Note:** Version bump only for package @nofrills/projector
+
+
+
+
+
+## [4.8.11](https://github.com/nativecode-dev/nofrills/compare/@nofrills/projector@4.8.10...@nofrills/projector@4.8.11) (2019-06-22)
+
+**Note:** Version bump only for package @nofrills/projector
+
+
+
+
+
 ## [4.8.10](https://github.com/nativecode-dev/nofrills/compare/@nofrills/projector@4.8.7...@nofrills/projector@4.8.10) (2019-06-22)
 
 **Note:** Version bump only for package @nofrills/projector